requirements.txt
Django>=5.0
argon2-cffi>=23.1
django-axes>=6.0
django-ratelimit>=3.1
django-csp>=3.8
cryptography>=42.0